% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stecf_landings_per_rectangle.R
\name{get_stecf_landings_per_rectangle}
\alias{get_stecf_landings_per_rectangle}
\title{Get and convert STECF landings and effort per rectangle data}
\usage{
get_stecf_landings_per_rectangle(file, nose_only, deep, fdf, format_long)
}
\arguments{
\item{file}{Define location where STECF landings per rectangle data is stored
as a standard csv file, as downloaded from http://stecf.jrc.ec.europa.eu/c/
document_library/get_file?uuid=015650e2-8d27-4c69-922c-2a15ba4d8b5a&groupId=43805}

\item{nose_only}{A logical indicating if you want all data or data from the
North Sea (ICES area IV) only (which is STECF reg gear code 3B in annex IIA).}

\item{deep}{A logical indicating if you want data of SPECON 'DEEP' to be
included. Standard value should be 'F' to avoid doubling data occurence.}

\item{fdf}{A logical indicating if you want data of 'fully documented
fisheries' to be included, such as 'FDFBAL', 'FDFIIA' and 'FDFIIC'.Standard
value should be 'F' to avoid doubling data occurence.}

\item{format_long}{A logical defining if you want the data to be transposed
to long format. Is do, specify \code{format_long <- 'T'}.}
}
\description{
\code{get_stecf_landings_per_rectangle} takes the downloaded STEFC data and
converts it to long format. Uppon request, it can subset data for the North
Sea only and remove data collected under the 'DEEP' specification.
}

