% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_in_map.R
\name{plot_factor_in_map}
\alias{plot_factor_in_map}
\title{Plot the value of a factor into a map of the North Sea}
\usage{
plot_factor_in_map(data, parameter, ices_rectangles, visuals = "colour")
}
\arguments{
\item{data}{Dataset to use for the plotting.}

\item{parameter}{Name of the parameters that are to be plot.}

\item{ices_rectangles}{Name of the column where location as ICES rectangles
are stored in the dataset.}

\item{visual}{A character string defining how you want the value of the
parameter to be represented. Use \code{visual = 'circles'} for point
size \code{cex} to represent values, use \code{visual = 'colour'} for
colour to represent values. Default is \code{visual = 'colour'}.}
}
\description{
\code{plot_factor_in_map} generates a map of the North Sea and plots the
value of a factor for each rectangle, either as point size or colour. The
function uses ICES rectangles to locate the data, conversion to coordinates
is provided.
}

